Added tower control mode and fixed serial delay
Added a 20ms timeout (instead of the default 1000ms) for serial.parseInt
calls.  Now the serial comm doesn't noticeably slow down the ring, so
you can smoothly adjust settings.  Also added the SomeoneIsInTheTower()
mode that will allow us to draw a pixel manually to the ring, hopefully
to be used for messing with people from the top of the effigy.